+++ Models/PartitionKeyStatistic.cs
@@ -5,6 +5,6 @@ namespace CosmosUtility.Models
     public class PartitionKeyStatistic
     {
         public List<string> PartitionKey { get; set; }
-        public int SizeInKB { get; set; }
+        public long SizeInKB { get; set; }
     }
 }
